--- a/build.gradle
+++ b/build.gradle
@@ -2,6 +2,7 @@ plugins {
 	id 'java'
 	id 'jacoco'
 	id 'com.diffplug.spotless' version '6.21.0'
+	id "me.champeau.jmh" version "0.7.2"
 }
 
 repositories {
@@ -34,6 +35,8 @@ test {
 
 dependencies {
 	testImplementation("org.junit.jupiter:junit-jupiter:5.7.2")
+	implementation 'org.openjdk.jmh:jmh-core:1.33'
+	implementation 'org.openjdk.jmh:jmh-generator-annprocess:1.33'
 }
 
 jacocoTestReport {