--- a/docs/benchmark.md
+++ b/docs/benchmark.md
@@ -4,9 +4,20 @@ Benchmarks can be ran with `./gradlew jmh`
 
 
 # Result of 07.01.2024 - Using arrays instead of HashSets
+
+## Before 
+```
+PerftBenchmark.position_2kr3r     thrpt    5  10.959 ± 0.117  ops/s
+PerftBenchmark.position_6bR       thrpt    5   5.637 ± 0.151  ops/s
+PerftBenchmark.position_r1b1k1r1  thrpt    5   3.728 ± 0.065  ops/s
+PerftBenchmark.position_r1bQk2r   thrpt    5  69.888 ± 4.363  ops/s
+```
+
+## After
 ```
 Benchmark                          Mode  Cnt   Score   Error  Units
-PerftBenchmark.position_2kr3r     thrpt    5  10.773 ± 0.535  ops/s
-PerftBenchmark.position_r1b1k1r1  thrpt    5   3.971 ± 0.074  ops/s
-PerftBenchmark.position_r1bQk2r   thrpt    5  88.351 ± 2.123  ops/s
+PerftBenchmark.position_2kr3r     thrpt    5  10.853 ± 0.335  ops/s
+PerftBenchmark.position_6bR       thrpt    5   5.647 ± 0.079  ops/s
+PerftBenchmark.position_r1b1k1r1  thrpt    5   3.990 ± 0.067  ops/s
+PerftBenchmark.position_r1bQk2r   thrpt    5  88.009 ± 4.822  ops/s
 ```