--- a/src/main/java/com/github/zerkath/rosemary/MoveGeneration/Knight.java
+++ b/src/main/java/com/github/zerkath/rosemary/MoveGeneration/Knight.java
@@ -2,11 +2,10 @@ package com.github.zerkath.rosemary.MoveGeneration;
 
 import com.github.zerkath.rosemary.BoardRepresentation.BoardState;
 import com.github.zerkath.rosemary.DataTypes.*;
-import java.util.HashMap;
 
 public class Knight {
 
-  static HashMap<Short, Moves> knightMoves = new HashMap<>();
+  static Moves[] knightMoves = new Moves[64];
 
   static {
     for (short origin = 0; origin < 64; origin++) {
@@ -25,14 +24,14 @@ public class Knight {
         Utils.addToCollection(d_row, col + 1, row, col, moves);
         Utils.addToCollection(d_row, col - 1, row, col, moves);
       }
-      knightMoves.put(origin, moves);
+      knightMoves[origin] = moves;
     }
   }
 
   public static void getMoves(short origin, BoardState boardState, Moves moves) {
 
     Board board = boardState.board;
-    Moves n_moves = knightMoves.get(origin);
+    Moves n_moves = knightMoves[origin];
 
     for (short move : n_moves) {
       if (board.isValidMove(move)) moves.add(move);